Move UnknownVersionException into version package, fix references and calls to exception
