+++ noora/plugins/mysql/update/UpdatePlugin.py
@@ -3,6 +3,7 @@ import os
 from noora.version.Version import Version
 from noora.version.Versions import Versions
 from noora.version.VersionLoader import VersionLoader
+from noora.version.UnknownVersionException import UnknownVersionException
 
 from noora.system.Ora import Ora
 from noora.system.PropertyHelper import PropertyHelper
@@ -10,7 +11,6 @@ from noora.io.File import File
 
 from noora.plugins.Plugin import Plugin
 from noora.plugins.Fail import Fail
-from noora.plugins.mysql.update.UnknownVersionException import UnknownVersionException
 from noora.plugins.mysql.update.InvalidEnvironmentException import InvalidEnvironmentException
 from noora.plugins.mysql.update.InvalidVersionException import InvalidVersionException
 
@@ -35,7 +35,7 @@ class UpdatePlugin(Plugin):
         alter_dir = properties.get('alter.dir')
         version_dir = File(os.path.join(alter_dir, version))
         if not version_dir.exists():
-            raise UnknownVersionException("unknown version folder", version)
+            raise UnknownVersionException("unknown version folder: {}".format(version))
 
     def fail_on_invalid_environment(self, connector, executor, environment, properties):
         plugin_dir = properties.get('plugin.dir')
